Implement projects in config, reshuffle code
I need to make timerFromResponseBody a method (not a function) because
it needs access to the config so it can just fill in the project as we
go. In perl toggl, I wrote `$toggl->oneline_desc($timer)`, but here I
think I'd rather just write `timer.OnelineDesc()` and not need to call
it on the singleton.