Implement 'resume' and associated methods
In particular, we needed a method to get a list of entries within a time
range (which we'll use in a bunch of other reporting commands too). This
required some minor refactoring in timer.go, since I wanted to reuse the
JSON extractor for multiple time entries, instead of just a single
entry.